narrow user deletion

+++ azure-integration/infrastructure/scripts/tear-down-azure.ps1
@@ -16,7 +16,10 @@ foreach ($subscription in $subscriptions) {
     Write-Host "getting role assignments for subscription '$($subscription.Id)'."
     $roleAssignments = Get-AzRoleAssignment -Scope $subscription.Id
     foreach ($roleAssignment in $roleAssignments) {
-        Remove-AzADUser -ObjectId $roleAssignment.ObjectId
-        Start-Sleep -Seconds 15
+        if ($roleAssignment.SignInName -like "*user*" -and $roleAssignment.ObjectType -eq "User") {
+            Write-Host "removing user '$($roleAssignment.SignInName)'"
+            Remove-AzADUser -ObjectId $roleAssignment.ObjectId
+            Start-Sleep -Seconds 15
+        }
     }
 }
